Add new class "Product"
I added a new class called 'product'. This class represents the product. Each product has unique ID and that is the primary key. The product also need to have title, description, price, owner email and in order to give more info about that product, we also give the last modified date of the product.