Modify the feature "Create product"4️⃣
There were several errors in the last changes commit.  I added the use of regular expression to filter if the title of the product has to be consists of alphabet, numbers and space. I also changed the code style this time to fit PEP8.  I also replaced 'return False' with 'return None'. There are also some misspell errors which have already been corrected.