Create test cases for create_product
There are three test parts in Create_product test which are input test, output test and function test.
test_create_product1 represents  the test of input, test all possibilities of input(one test case for each categories)
test_create_product2 represents  the test of output, test all possibilities of out(one test case for each categories)
test_create_product1 represents  the test of function, to test the function, we need to make sure that the product has been created and added to the database successfully. To test that, we can try to update the product information to see if the system can find the product.